[CI/Build] Extract shared utilities from tests/conftest.py into tests/tools/
The root conftest.py (3,180 lines) served as both a pytest fixture file
and an importable utility library. This extracts all non-fixture code
into tests/tools/ modules while keeping backward compatibility via
re-exports.

New modules:
- tests/tools/types.py — OmniServerParams, OmniResponse, DiffusionResponse
- tests/tools/assertions.py — all assert_* validation functions
- tests/tools/text_processing.py — preprocess_text, cosine_similarity_text
- tests/tools/voice_analysis.py — gender detection, pitch, HNR analysis
- tests/tools/audio_processing.py — Whisper transcription helpers
- tests/tools/media_generators.py — synthetic audio/video/image generators
- tests/tools/config_utils.py — modify_stage_config YAML helper
- tests/tools/gpu_cleanup.py — GPU memory cleanup utilities
- tests/tools/server.py — OmniServer class
- tests/tools/client.py — OpenAIClientHandler class
- tests/tools/runner.py — OmniRunner, OmniRunnerHandler classes

conftest.py now contains only fixtures and re-exports (~200 lines).
Existing `from tests.conftest import X` statements continue to work.

Step A of RFC #2299.
